feat: enhance CI test visibility and fix remaining Windows tests
CI Improvements:
- Add JSON test output and test result artifacts
- Integrate go-junit-report for JUnit XML generation
- Add dorny/test-reporter for better GitHub UI integration
- Create custom test summary script with visual progress bars
- Add GitHub Actions problem matchers for Go test output
- Update permissions to allow checks and PR comments
- Add Makefile targets for local test summaries (test-json, test-summary, test-watch)

Test Fixes:
- Fix TestInitCaches on Windows by using NUL device path
- Fix SaveProjectConfig validation to check directory containment properly
- Move Unix-specific permission tests to config_test_unix.go with build tags
- Update error message expectations for cross-platform compatibility
- Fix os.Getuid() compilation error on Windows

These changes provide much better visibility into test failures in GitHub Actions,
making it easier to identify and fix issues without digging through raw logs.
