+++ .github/workflows/ci.yml
@@ -8,6 +8,8 @@ on:
 
 permissions:
   contents: read
+  checks: write
+  pull-requests: write
 
 jobs:
   test:
@@ -30,9 +32,67 @@ jobs:
     - name: Get dependencies
       run: go mod download
     
-    - name: Run tests
+    - name: Setup problem matcher
+      run: echo "::add-matcher::.github/problem-matchers/go-test.json"
+    
+    - name: Run tests with JSON output
       run: |
-        go test -v -race -coverprofile coverage.txt -covermode atomic ./...
+        go test -v -race -coverprofile coverage.txt -covermode atomic -json ./... > test-results.json || true
+        
+    - name: Generate test report
+      if: always()
+      run: |
+        go install github.com/jstemmer/go-junit-report/v2@latest
+        cat test-results.json | go-junit-report -parser gojson > test-results.xml
+    
+    - name: Create test summary
+      if: always()
+      run: |
+        # Run the custom test summary script
+        bash .github/scripts/test-summary.sh test-results.json || true
+        
+        # Also run gotestsum for additional output
+        go install github.com/dnephin/gotestsum@latest
+        gotestsum --jsonfile test-results.json --format dots-v2 || true
+    
+    - name: Upload test results
+      if: always()
+      uses: actions/upload-artifact@v4
+      with:
+        name: test-results-${{ matrix.os }}-go${{ matrix.go }}
+        path: |
+          test-results.json
+          test-results.xml
+    
+    - name: Publish test results
+      if: always()
+      uses: dorny/test-reporter@v1
+      with:
+        name: Test Results (${{ matrix.os }} - Go ${{ matrix.go }})
+        path: test-results.xml
+        reporter: java-junit
+        fail-on-error: false
+        
+    - name: Comment PR with test results
+      if: always() && github.event_name == 'pull_request' && failure()
+      uses: actions/github-script@v7
+      with:
+        github-token: ${{secrets.GITHUB_TOKEN}}
+        script: |
+          const fs = require('fs').promises;
+          try {
+            const summary = await fs.readFile(process.env.GITHUB_STEP_SUMMARY, 'utf8');
+            const comment = `### Test Results for ${{ matrix.os }} - Go ${{ matrix.go }}\n\n${summary}`;
+            
+            github.rest.issues.createComment({
+              issue_number: context.issue.number,
+              owner: context.repo.owner,
+              repo: context.repo.repo,
+              body: comment
+            });
+          } catch (error) {
+            console.log('Could not read test summary:', error);
+          }
     
     - name: Upload coverage
       if: matrix.os == 'ubuntu-latest' && matrix.go == '1.24'
